requirements.txt
streamlit==1.28.0
langchain==0.1.0
langchain-groq==0.1.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
duckduckgo-search==4.4.0